rename api to backend, app to frontend and chatbot to core
