+++ deploy_chatbot_python/backend/run.py
@@ -1,6 +1,6 @@
 import uvicorn
 
-from deploy_chatbot_python.api.server import api  # pylint: disable=unused-import
+from deploy_chatbot_python.backend.server import api  # pylint: disable=unused-import
 from deploy_chatbot_python.config import constants
 
 if __name__ == "__main__":
